Bot-Telegram/requirements.txt
python-telegram-bot==21.*
anthropic
python-dotenv
